refactor: move types to separated file
